refactor: rename process group
