+++ paddle/fluid/distributed/collective/custom_ccl_tools.cc
@@ -12,8 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "paddle/fluid/distributed/collective/CustomCCLTools.h"
-#include "paddle/fluid/distributed/collective/Types.h"
+#include "paddle/fluid/distributed/collective/custom_ccl_tools.h"
+#include "paddle/fluid/distributed/collective/types.h"
 
 namespace paddle {
 namespace distributed {
