+++ paddle/fluid/distributed/collective/process_group.h
@@ -19,7 +19,7 @@
 #include <string>
 #include <vector>
 
-#include "paddle/fluid/distributed/collective/Types.h"
+#include "paddle/fluid/distributed/collective/types.h"
 #include "paddle/fluid/eager/api/utils/tensor_utils.h"
 #include "paddle/fluid/framework/tensor.h"
 #include "paddle/fluid/framework/variable.h"
