+++ paddle/phi/kernels/gpu/class_center_sample_kernel.cu
@@ -33,7 +33,7 @@ namespace cub = hipcub;
 #include "paddle/phi/core/enforce.h"
 
 #if defined(PADDLE_WITH_NCCL) || defined(PADDLE_WITH_RCCL)
-#include "paddle/fluid/distributed/collective/ProcessGroup.h"
+#include "paddle/fluid/distributed/collective/process_group.h"
 #include "paddle/fluid/platform/collective_helper.h"
 #include "paddle/fluid/platform/device/gpu/nccl_helper.h"
 #endif
